Moves functions folder to outside of app folder
