--- app/functions/create-board.ts
@@ -1,22 +1,0 @@
-"use server";
-
-import { db } from "@/prisma/db";
-import { z } from "zod";
-
-const CreateBoard = z.object({
-  title: z.string().min(3, {
-    message: "Minimum length of 3 letters is required",
-  }),
-});
-
-export async function create(formData: FormData) {
-  const { title } = CreateBoard.parse({
-    title: formData.get("title"),
-  });
-
-  await db.board.create({
-    data: {
-      title,
-    },
-  });
-}
