+++ functions/create-board.ts
@@ -1,0 +1,37 @@
+"use server";
+
+import { db } from "@/prisma/db";
+import { z } from "zod";
+
+const CreateBoard = z.object({
+  title: z.string().min(3, {
+    message: "Minimum length of 3 letters is required",
+  }),
+});
+
+export async function create(formData: FormData) {
+  const validatedFields = CreateBoard.safeParse({
+    title: formData.get("title"),
+  });
+
+  if (!validatedFields.success) {
+    return {
+      errors: validatedFields.error.flatten().fieldErrors,
+      message: "missing fields",
+    };
+  }
+
+  const { title } = validatedFields.data;
+
+  try {
+    await db.board.create({
+      data: {
+        title,
+      },
+    });
+  } catch (error) {
+    return {
+      message: "Database Error",
+    };
+  }
+}
